requirements.txt
streamlit>=1.30.0
requests>=2.31.0
openpyxl>=3.1.2
apscheduler>=3.10.4
pyyaml>=6.0.1
python-dotenv>=1.0.0
pdfplumber>=0.10.0
supabase>=2.0.0
